Change completion install directory
The bash-completion/completions directory is meant for lazy loading of
completion scripts by matching the command with the completion script
name I believe.

If I got this this correct this wouldn't work for git sub commands and
the note to load all completions from
~/.local/share/bash-completion/completions would defeat the lazy loading
anyways. So just install into another random directory and ask the user
to use that instead.

Oh I wish there was proper standards for user installs of
bash completion.